workers/runpod-comfyui/requirements.txt
runpod==1.7.9
requests==2.32.3
websocket-client==1.8.0
boto3==1.35.90

